api2/requirements.txt
fastapi
paho-mqtt
